Add some comments, move things around
